Removed imports from 'internal.ts', removed unused 'NotificationsEvents' type alias. Removed 'openfin' types from test tsconfig.json.
